+++ opencti-platform/opencti-front/src/private/components/TimeoutLock.tsx
@@ -8,7 +8,7 @@ import { useFormatter } from '../../components/i18n';
 import { APP_BASE_PATH, fetchQuery } from '../../relay/environment';
 import { formatSeconds, ONE_SECOND, secondsBetweenDates } from '../../utils/Time';
 import useAuth from '../../utils/hooks/useAuth';
-import useTopBannerHeight from '../../utils/hooks/useTopBannerHeight';
+import useTopBanner from '../../utils/hooks/useTopBanner';
 
 /**
  * Gets timeout and banner settings from react relay and return those values.
@@ -78,7 +78,7 @@ const TimeoutLock: React.FunctionComponent = () => {
   const {
     bannerSettings: { bannerHeightNumber, idleLimit, sessionLimit },
   } = useAuth();
-  const topBannerHeight = useTopBannerHeight();
+  const { height: topBannerHeight } = useTopBanner();
   const [dialogOpen, setDialogOpen] = useState(false);
   const [state, dispatch] = useReducer(timeoutReducer, {
     idleLimit,
