+++ opencti-platform/opencti-front/src/private/components/TopBannersManager.tsx
@@ -1,29 +1,22 @@
-import React, { useContext, useState } from 'react';
-import { UserContext } from '../../utils/hooks/useAuth';
-import useGranted, { SETTINGS_SETMANAGEXTMHUB } from '../../utils/hooks/useGranted';
-import useHelper from '../../utils/hooks/useHelper';
-import { REGISTER_BANNER_DISMISSED_BUS, readRegisterDismissed, shouldDisplayLicenseBanner, shouldDisplayRegisterBanner, shouldDisplayTrialBanner } from '../../utils/bannerUtils';
+import React, { useEffect } from 'react';
+import useTopBanner from '../../utils/hooks/useTopBanner';
 import LicenseBanner from './LicenseBanner';
 import StartTrialBanner from './xtm_hub/StartTrialBanner';
 import RegisterPlatformBanner from './xtm_hub/RegisterPlatformBanner';
-import useBus from '../../utils/hooks/useBus';
+import useAuth from '../../utils/hooks/useAuth';
+import { resetRegisterBannerDismiss } from '../../utils/bannerUtils';
 
 const TopBannersManager = () => {
-  const { settings, isXTMHubAccessible } = useContext(UserContext);
-  const isGrantedToXtmHub = useGranted([SETTINGS_SETMANAGEXTMHUB]);
-  const { isFeatureEnable } = useHelper();
+  const { showLicenseBanner, showTrialBanner, showRegisterBanner } = useTopBanner();
+  const { settings } = useAuth();
 
-  const [isDismissed, setIsDismissed] = useState<boolean>(readRegisterDismissed);
-  useBus(REGISTER_BANNER_DISMISSED_BUS, setIsDismissed, []);
-
-  const showLicenseBanner = shouldDisplayLicenseBanner(settings?.platform_enterprise_edition);
-  const showTrialBanner = shouldDisplayTrialBanner(settings);
-  const showRegisterBanner = isFeatureEnable('XTMHUB_NEWS_FEED_ENABLED') && shouldDisplayRegisterBanner(
-    settings,
-    isXTMHubAccessible,
-    isGrantedToXtmHub,
-    isDismissed,
-  );
+  // Auto-reset dismissed state once the platform becomes registered,
+  // so the banner reappears if the platform later becomes unregistered.
+  useEffect(() => {
+    if (settings?.xtm_hub_registration_status === 'registered') {
+      resetRegisterBannerDismiss();
+    }
+  }, [settings?.xtm_hub_registration_status]);
 
   return (
     <>
