+++ opencti-platform/opencti-front/src/utils/bannerConstants.ts
@@ -1,0 +1,3 @@
+export const REGISTER_BANNER_DISMISSED_KEY = 'register-banner-dismissed';
+export const REGISTER_BANNER_DISMISSED_BUS = `${REGISTER_BANNER_DISMISSED_KEY}_bus`;
+
